Moved custom hook to a separate folder
